requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-firestore==1.2.0
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.3.1
pytz==2019.1
requests==2.22.0
robobrowser==0.5.3
rsa==4.0
sentry-sdk==0.8.0
six==1.12.0
soupsieve==1.9.1
typed-ast==1.3.5
urllib3==1.25.2
Werkzeug==0.15.4
wrapt==1.11.1
